fix(src): resolve #18 reviews
